ai-backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-core
langchain-ollama
colorlog
python-multipart
pymupdf
unstructured
langchain-community
faiss-cpu
sentence-transformers
python-dotenv
qdrant-client
minio
WebSearcher
fastapi-mcp
duckduckgo-search
bs4
langgraph